@@ -8,30 +8,25 @@
     "watch": "webpack --mode=development --watch"
   },
   "dependencies": {
-    "@bitcoin-design/bitcoin-icons-svg": "^0.1.9",
-    "@emotion/react": "^11.8.1",
-    "@figma/plugin-typings": "^1.42.1",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@bitcoin-design/bitcoin-icons-svg": "^0.1.10",
+    "@emotion/react": "^11.9.0",
+    "@figma/plugin-typings": "^1.44.0",
+    "@types/react": "^18.0.6",
+    "@types/react-dom": "^18.0.2",
     "@types/react-toggle": "^4.0.3",
-    "css-loader": "^6.6.0",
+    "css-loader": "^6.7.1",
     "fuse.js": "^6.5.3",
     "html-webpack-plugin": "^5.5.0",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dev-utils": "^12.0.0",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dev-utils": "^12.0.1",
+    "react-dom": "^18.0.0",
     "react-toggle": "^4.1.2",
     "style-loader": "^3.3.1",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.5.5",
-    "webpack": "^5.69.1",
+    "ts-loader": "^9.2.8",
+    "typescript": "^4.6.3",
+    "webpack": "^5.72.0",
     "webpack-cli": "^4.9.2"
   },
-  "packageManager": "yarn@3.2.0",
-  "devDependencies": {
-    "@rollup/plugin-typescript": "^8.3.1",
-    "rollup": "^2.68.0",
-    "rollup-plugin-inline-js": "^0.5.0"
-  }
+  "packageManager": "yarn@3.2.0"
 }
